Have menu set up by UIStarter, not ResultsTable
All this functionality used to be in UIStarter (and ResultsTable didn't exist).  I created ResultsTable to break out the functionality that was specific to the table.  I should not have included the menu generation; that is more generic and better fits in UIStarter.  Now UIStarter does things like initialize the Cocoa functionality, which makes a lot more sense than having a class that represents a Table do it.
